+++ client/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
+import App from './ChatScreen';
 import { Provider } from 'react-redux';
 import store from './redux/store';
 import { AppThemeProvider } from './services/AppThemeProvider';
